requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
cffi==1.14.6
cycler==0.10.0
decorator==5.1.0
face-alignment==1.3.5
imageio==2.9.0
imageio-ffmpeg==0.4.5
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.22.4
pandas==1.3.3
Pillow==8.2.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
tensorboardx
tqdm==4.62.3
pytorch_fid_wrapper
torchjpeg==0.9.23
torch==1.12.1+cu113
torchvision==0.13.1+cu113
opencv-python==4.6.0.66
lpips